pyproject.toml
[tool.poetry]
name = "femsntl"
version = "0.1.0"
description = "Code for analyzing the NTL Project"
authors = ["The Lab at DC <thelab@dc.gov>"]
packages = [
    { include = "femsntl", from = "src" }
]
license = "CC-0"

[tool.poetry.scripts]
ntl = "femsntl.cli:cli"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.19.2"
pandas = "1.1.2"
jupyter = "^1.0.0"
xlrd = "^1.2.0"
python-dateutil = "^2.8.1"
statsmodels = "^0.12.0"
pymssql = "^2.1.5"
pyyaml = "^5.3.1"
click = "^8.0.1"
plotnine = "^0.8.0"
recordlinkage = "^0.14"
gender-guesser = "^0.4.0"
fuzzywuzzy = "^0.18.0"
lxml = "^4.6.3"
openpyxl = "^2.6.4"
python-Levenshtein = "^0.12.2"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.0.2"
pylint = "^2.6.0"
isort = "^5.7.0"
ipdb = "^0.13.4"
mypy = "^0.910"
types-requests = "^2.25.6"
pytest-postgresql = "^3.1.1"
nbqa = "^1.1.1"
papermill = "^2.3.3"
types-PyYAML = "^5.4.10"
nbstripout = "^0.5.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.basic]
good-names = "i,j,k,ex,Run,_,df,pc"

[tool.pylint.messages_control]
disable = "C0330, C0326, R0912, R0913, R0914, R0915"

[tool.pylint.format]
max-line-length = "88"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
